pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "pulsar"
version = "0.1.0"
description = "Smart driver distribution runtime and API"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Pulsar Team" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.0",
  "redis>=5.0.1",
  "aio-pika>=9.3.1",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "h3>=4.1.0",
  "scikit-learn>=1.4.0",
  "python-dateutil>=2.8.2",
  "pyarrow>=17.0.0",
  "jinja2>=3.1.0",
  "mlflow>=2.17.0",
  "torch>=2.4.0",
  "pika>=1.3.2",
  "nats-py>=2.7.0",
  "clickhouse-connect>=0.8.0",
]

[project.scripts]
pulsar = "pulsar.cli:main"

[tool.hatch.build.targets.wheel]
packages = [
  "src/pulsar",
  "src/pulsar_core",
]

[tool.hatch.build]
include = [
  "src/pulsar_core/templates/forecast.html",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.3",
]
